requirements.txt

flask
gtts
ffmpeg-python
